app/requirements.txt
flask
google-genai
gunicorn  
